Created fragments for profile, feed, and explore, connected bottom navigation view to fragments, transfer related code from MainActivity to Feed fragment
