+++ app/src/main/java/com/example/event_finder_capstone/activities/MainActivity.java
@@ -8,127 +8,53 @@ import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
 
-import com.example.event_finder_capstone.adapters.EventsAdapter;
 import com.example.event_finder_capstone.R;
-import com.example.event_finder_capstone.network.RetrofitClient;
-import com.example.event_finder_capstone.models.Event;
-import com.google.gson.JsonArray;
-import com.google.gson.JsonObject;
+import com.example.event_finder_capstone.fragments.ExploreFragment;
+import com.example.event_finder_capstone.fragments.FeedFragment;
+import com.example.event_finder_capstone.fragments.ProfileFragment;
+import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.parse.ParseUser;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
-
 public class MainActivity extends AppCompatActivity {
-    private static final String TAG = "MainActivity";
-    RecyclerView rvEvents;
-    private List<Event> eventsList = new ArrayList<>();
-    private EventsAdapter eventsAdapter;
+
+    private BottomNavigationView bottomNavigationView;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        setUpRecyclerView();
-        getAPIEvents();
+        final FragmentManager fragmentManager = getSupportFragmentManager();
+        setUpBottomNavigation(fragmentManager);
     }
 
-    private void setUpRecyclerView() {
-        rvEvents = findViewById(R.id.rvEvents);
-        eventsList = new ArrayList<>();
-        eventsAdapter = new EventsAdapter(this, eventsList);
-        rvEvents.setAdapter(eventsAdapter);
-        LinearLayoutManager linearLayoutManager = new LinearLayoutManager(this);
-        rvEvents.setLayoutManager(linearLayoutManager);
-    }
+    private void setUpBottomNavigation(FragmentManager fragmentManager) {
+        bottomNavigationView = findViewById(R.id.bottom_navigation);
 
-    private void getAPIEvents() {
-        String eventSearchRegion = "en_US";
-        Long eventSearchRadiusFromUserInMeters = 40000L;
-        String numberOfEventsToRetrieve = "10";
-        Long upcomingEventsOnly = (System.currentTimeMillis() / 1000L);
-        RetrofitClient.getInstance().getYelpAPI().getEvents(eventSearchRegion,
-                numberOfEventsToRetrieve,
-                upcomingEventsOnly,
-                eventSearchRadiusFromUserInMeters,
-                ParseUser.getCurrentUser().getString("zip")).enqueue(new Callback<JsonObject>() {
+        bottomNavigationView.setOnItemSelectedListener(new BottomNavigationView.OnItemSelectedListener() {
             @Override
-            public void onResponse(@NonNull Call<JsonObject> call, @NonNull Response<JsonObject> response) {
-                if (response.isSuccessful() && response.body() != null) {
-                    runOnUiThread(() -> {
-                        try {
-                            JsonObject result = response.body();
-                            eventsList.clear();
-                            eventsList.addAll(convertToList(result));
-                            eventsAdapter.notifyDataSetChanged();
-                        } catch (Exception e) {
-                            Toast.makeText(MainActivity.this, "JSON exception error", Toast.LENGTH_SHORT).show();
-                        }
-                    });
-                } else {
-                    Toast.makeText(MainActivity.this, "Query Failed", Toast.LENGTH_SHORT).show();
+            public boolean onNavigationItemSelected(@NonNull MenuItem menuItem) {
+                Fragment fragment;
+                switch (menuItem.getItemId()) {
+                    case R.id.action_feed:
+                        fragment = new FeedFragment();
+                        break;
+                    case R.id.action_explore:
+                        fragment = new ExploreFragment();
+                        break;
+                    case R.id.action_profile:
+                        fragment = new ProfileFragment();
+                        break;
+                    default:
+                        return true;
                 }
-            }
-
-            @Override
-            public void onFailure(@NonNull Call<JsonObject> call, @NonNull Throwable t) {
-                Toast.makeText(MainActivity.this, "Failed to call API", Toast.LENGTH_SHORT).show();
+                fragmentManager.beginTransaction().replace(R.id.flContainer, fragment).commit();
+                return true;
             }
         });
-    }
-
-    private Collection<? extends Event> convertToList(JsonObject result) {
-        List<Event> res = new ArrayList<>();
-        JsonArray events = result.getAsJsonArray("events");
-        for (int i = 0; i < events.size(); i++) {
-            JsonObject temp = (JsonObject) events.get(i);
-            Event event = new Event();
-            populateEventInfo(i, event, temp);
-            res.add(event);
-        }
-        return res;
-    }
-
-    private void populateEventInfo(int i, Event event, JsonObject temp) {
-        event.setName(temp.get("name").getAsString());
-        event.setDescription(temp.get("description").getAsString());
-        event.setImageUrl(temp.get("image_url").getAsString());
-        event.setTimeStart(temp.get("time_start").getAsString());
-        event.setId(temp.get("id").getAsString());
-        event.setEventSiteUrl(temp.get("event_site_url").getAsString());
-        checkAndSetEventEndTime(event, temp);
-        checkAndSetEventCost(event, temp);
-        formatAndSetEventLocation(temp, event);
-    }
-
-    private void checkAndSetEventCost(Event event, JsonObject temp) {
-        if (!String.valueOf(temp.get("cost")).equals("null")) {
-            event.setCost("$" + temp.get("cost").getAsString() + "0");
-        } else {
-            event.setCost("N/A");
-        }
-    }
-
-    private void checkAndSetEventEndTime(Event event, JsonObject temp) {
-        if (!String.valueOf(temp.get("time_end")).equals("null")) {
-            event.setTimeEnd(temp.get("time_end").getAsString());
-        } else {
-            event.setTimeEnd(temp.get("time_start").getAsString());
-        }
-    }
-
-    private void formatAndSetEventLocation(JsonObject temp, Event event) {
-        String formattedLocation = temp.get("location").getAsJsonObject().get("display_address").toString();
-        formattedLocation = formattedLocation.replaceAll("[\\[\\]\"\"]", " ");
-        event.setLocation(formattedLocation);
+        bottomNavigationView.setSelectedItemId(R.id.action_feed);
     }
 
     @Override
