+++ app/src/main/res/menu/menu_bottom_navigation.xml
@@ -4,17 +4,21 @@
 
     <item
         android:id="@+id/action_profile"
+        android:enabled="true"
         android:icon="@drawable/ic_baseline_person_24"
         android:title="@string/profile"
-        app:showAsAction="always" />
+        app:showAsAction="ifRoom" />
     <item
         android:id="@+id/action_feed"
+        android:enabled="true"
         android:icon="@android:drawable/ic_menu_compass"
         android:title="@string/feed"
         app:showAsAction="ifRoom" />
     <item
         android:id="@+id/action_explore"
+        android:enabled="true"
         android:icon="@drawable/ic_baseline_search_24"
         android:title="@string/explore"
-        app:showAsAction="always" />
+        app:showAsAction="ifRoom"
+        />
 </menu>
